+++ Libplanet/Net/Swarm.cs
@@ -220,6 +220,9 @@ namespace Libplanet.Net
             CancellationToken cancellationToken = default(CancellationToken)
         )
         {
+            _logger.Debug("Stopping watching " + nameof(BlockChain) + " for tip changes...");
+            BlockChain.TipChanged -= OnBlockChainTipChanged;
+
             _logger.Debug($"Stopping {nameof(Swarm<T>)}...");
             using (await _runningMutex.LockAsync())
             {
@@ -321,6 +324,9 @@ namespace Libplanet.Net
 
             _logger.Debug("Starting swarm...");
             _logger.Debug("Peer information : {Peer}", AsPeer);
+
+            _logger.Debug("Watching the " + nameof(BlockChain) + " for tip changes...");
+            BlockChain.TipChanged += OnBlockChainTipChanged;
 
             try
             {
@@ -440,6 +446,15 @@ namespace Libplanet.Net
             }
         }
 
+        /// <summary>
+        /// Broadcasts the given block to peers.
+        /// <para>The message is immediately broadcasted, and it is done if the same block has
+        /// already been broadcasted before.</para>
+        /// </summary>
+        /// <param name="block">The block to broadcast to peers.</param>
+        /// <remarks>It does not have to be called manually, because <see cref="Swarm{T}"/> in
+        /// itself watches <see cref="BlockChain"/> for <see cref="BlockChain{T}.Tip"/> changes and
+        /// immediately broadcasts updates if anything changes.</remarks>
         public void BroadcastBlock(Block<T> block)
         {
             BroadcastBlock(null, block);